cleaned up lint excludes and improved redact string
