+++ .pre-commit-config.yaml
@@ -22,7 +22,7 @@ repos:
     -   id: mixed-line-ending
     -   id: trailing-whitespace
 -   repo: https://github.com/golangci/golangci-lint
-    rev: v1.56.2
+    rev: v1.57.1
     hooks:
     -   id: golangci-lint
 -   repo: local
